добавили сборки разнесли код по слоям добавили фильтры в гет запрос
