+++ EShop.Domain/Product.cs
